Properly define typescript types as module
