+++ fhir-context/stu3/index.d.ts
@@ -1,0 +1,8 @@
+import { Model } from "../../src/fhirpath";
+
+export const {
+    choiceTypePaths,
+    pathsDefinedElsewhere,
+    type2Parent,
+    path2Type
+}: Model;
