+++ package.json
@@ -3,6 +3,7 @@
   "version": "3.6.1",
   "description": "A FHIRPath engine",
   "main": "src/fhirpath.js",
+  "types": "src/fhirpath.d.ts",
   "dependencies": {
     "@lhncbc/ucum-lhc": "^4.1.3",
     "antlr4": "~4.9.3",
